Catchup update

+++ ttsController.py
@@ -3,7 +3,9 @@ from preferredsoundplayer import *
 from TTS.utils.synthesizer import Synthesizer
 from twitchAPI.oauth import UserAuthenticator, refresh_access_token
 from twitchAPI.twitch import Twitch
-from twitchAPI.types import AuthScope
+from twitchAPI.type import AuthScope
+
+import asyncio
 import configparser
 import emoji
 import glob
@@ -19,9 +21,7 @@ import string
 import threading
 import urllib.parse
 import websocket
-import pysynth_s as synth_t
-import pysynth_b as synth_b
-import mixfiles
+import winsdk.windows.media.control as wmc
 
 DEVMODE = False
 
@@ -46,6 +46,12 @@ def remove_cheermotes(raw_message):
         message += word + " "
     message.strip()
     return message
+
+
+async def get_media_session():
+    sessions = await wmc.GlobalSystemMediaTransportControlsSessionManager.request_async()
+    session = sessions.get_current_session()
+    return session
 
 
 class ttsController:
@@ -104,6 +110,7 @@ class ttsController:
         self.tts_thread = None 
 
         self.currently_playing = None
+        self.skip_flag = False
         self.pause_flag = False
         self.clear_flag = False
         self.speaker_list = eval(self.config['DEFAULT']['Speakers'])
@@ -132,6 +139,7 @@ class ttsController:
                     self.tts_thread = self.tts_queue.get(timeout=1)
                     self.tts_thread()
                 except (queue.Empty, IndexError):
+                    self.skip_flag = False
                     continue
             else:
                 sleep(1)
@@ -145,11 +153,13 @@ class ttsController:
         return fname
 
     def generate_wav(self, msg):
+        user = msg['user_name']
         for id_msg, message_object in enumerate(msg['message']):
             output_file = self.generate_fname()
             prepend_file = None
 
             if self.clear_flag: continue
+            print(message_object)
             voice = message_object['voice']
             message = message_object['message']
 
@@ -170,92 +180,13 @@ class ttsController:
             elif voice == 'lute':
                 # MUSIC
                 try:
-                    tempo = 140
-                    tempo_check = ' '.join(message.split()).split()
-                    if len(tempo_check) == 2:
-                        tempo = max(int(tempo_check[0]), 70)
-                        print(f'Setting the tempo to {tempo}')
-                        message = tempo_check[1]
-
-                    lines = message.split('|')
-                    line_lengths = []
-                    full_notation = []
-                    for id_line, line in enumerate(lines):
-                        beats = line.split('-')
-                        notation = []
-                        length = 1
-                        line_length = 0.
-                        repeat = False
-                        repeat_list = []
-                        for id_beat, beat in enumerate(beats):
-                            if beat == ':':
-                                if repeat:
-                                    for id_rep, rep in enumerate(repeat_list):
-                                        beats.insert(id_beat + 1 + id_rep, rep)
-                                    print(f'Repetition: ' + str(repeat_list))
-                                repeat = not repeat
-                                repeat_list = []
-                            else:
-                                if repeat:
-                                    repeat_list.append(beat)
-                                note = beat.split('.', 1)
-
-                                if ''.join(filter(str.isalpha, note[0])) in ttsController.NOTES:
-                                    str_note = note[-1]
-                                    if len(note) != 1:
-                                        if str_note[0] == '/':
-                                            str_note = str_note[1:]
-                                            if str_note[-1] == '*':
-                                                str_note = float(str_note[:-1]) / 1.5
-                                            length = 1 / float(str_note)
-
-                                        else:
-                                            if str_note[-1] == '*':
-                                                str_note = float(str_note[:-1]) * 1.5
-                                            length = float(str_note)
-                                    line_length += length
-                                    notation.append((note[0], 4. / length))
-
-                        line_lengths.append(line_length)
-                        full_notation.append(notation)
-
-                    max_length = max(line_lengths)
-                    last_file = self.generate_fname()
-                    for id_line, line in enumerate(full_notation):
-                        cur_file = self.generate_fname()
-
-                        if line_lengths[id_line] != max_length:
-                            print(f'Line too short, adding rest to match {line_lengths[id_line]} to {max_length}')
-                            line.append(('r', float(4. / (max_length - line_lengths[id_line]))))
-                            print(tuple(line))
-
-                        if line != full_notation[-1] or len(full_notation) == 1:
-                            synth_t.make_wav(tuple(line), fn=self.output_path + cur_file + '.wav', bpm=tempo)
-                        else:
-                            synth_b.make_wav(tuple(line), fn=self.output_path + cur_file + '.wav', bpm=tempo)
-
-                        amix = 0.5
-                        bmix = 0.5
-                        if id_line == 2: bmix = 1.
-                        if id_line >= 1 and line == full_notation[-1]: amix = 0.4
-                        if id_line > 0:
-                            new_file = self.generate_fname()
-                            print(f'Mixing files {cur_file} and {last_file} into {new_file}')
-                            mixfiles.mix_files(self.output_path + cur_file + '.wav', \
-                                                self.output_path + last_file + '.wav', \
-                                                self.output_path + new_file + '.wav', amix, bmix, 1)
-                            print(f'Mixed {new_file}')
-                            os.remove(self.output_path + cur_file + '.wav')
-                            os.remove(self.output_path + last_file + '.wav')
-                            last_file = new_file
-                        else:
-                            print(f'Tracking last file {cur_file}')
-                            last_file = cur_file
-
-                    message_object['message'] = '-'
-                    output_file = last_file
+                    url = f'https://luteboi.com/lute/?message={urllib.parse.quote_plus("#lute " + message)}&key={urllib.parse.quote_plus(user)}'
+                    data = requests.get(url)
+                    with open(self.output_path + output_file + '.wav', 'wb') as f:
+                        f.write(data.content)
                 except Exception as ex:
                     print(f'Generation done broke. Sorry luter: {ex}')
+                    shutil.copyfile('./assets/broken.wav', self.output_path + output_file + '.wav')
             else:
                 # Do Brian
                 url = 'https://api.streamelements.com/kappa/v2/speech?voice=Brian&text=' + urllib.parse.quote_plus(
@@ -263,7 +194,6 @@ class ttsController:
                 data = requests.get(url)
                 with open(self.output_path + output_file + '.mp3', 'wb') as f:
                     f.write(data.content)
-                f.close()
 
             # Put the file in the play queue
             print(f'New file for {voice}: {message}: {output_file}')
@@ -277,22 +207,34 @@ class ttsController:
     def play_wav(self, message_list):
         soundplay("assets/cheer.wav", block=True)
         sleep(0.25)
-
+        media_paused = False
+        session = None
         for message_object in message_list:
+            if message_object['voice'] == "lute":
+                session = asyncio.run(get_media_session())
+                if session is not None:
+                    media_paused = session.get_playback_info().playback_status.value == 4
+                    session.try_pause_async()
+
             file = None
             for f in os.listdir(self.output_path):
                 if message_object['filename'] == f[0:6]: file = self.output_path + f
             try:
-                self.currently_playing = soundplay(file)
-                while getIsPlaying(self.currently_playing):
-                    if self.clear_flag:
-                        stopsound(self.currently_playing)
-                    sleep(0.1)
-                stopsound(self.currently_playing)
-                os.remove(file)
+                if file:
+                    self.currently_playing = soundplay(file)
+                    while getIsPlaying(self.currently_playing):
+                        if self.clear_flag or self.skip_flag:
+                            stopsound(self.currently_playing)
+                            self.skip_flag = False
+                        sleep(0.1)
+                    stopsound(self.currently_playing)
+                    os.remove(file)
             except:
                 print(f'Could not play file.')
-                continue
+
+            if session is not None and media_paused:
+                session.try_play_async()
+                media_paused = False
 
         # Hack way to let GUI know to decrease the visible messages?
         self.currently_playing = None
@@ -343,7 +285,8 @@ class ttsController:
                         'type': sub_type,
                         'version': '1',
                         'condition': {
-                            'broadcaster_user_id': self.broadcaster['data'][0]['id']
+                            'broadcaster_user_id': self.broadcaster['data'][0]['id'],
+                            'user_id': self.broadcaster['data'][0]['id']
                        },
                         'transport': {
                             'method': 'websocket',
@@ -431,7 +374,6 @@ class ttsController:
 
                     if sub_message.strip() != '': message_list.append(sub_message_object)
 
-        print(message_list)
         return message_list
 
     async def update_stored_creds(self, token, refresh):
@@ -472,6 +414,8 @@ class ttsController:
                 broad_request = requests.get(f'{ttsController.URI}/users?user_login={self.target_channel}',
                                              headers=self.headers)
                 self.broadcaster = broad_request.json()
+
+                print(self.broadcaster)
 
         # Create the socket for threading
         if not self.wsapp:
